Fixed bugs in search functions
Prematurely appended and removed in search function. For example, I would remove/append to current in the second for loop and this was wron because it could change the value of currentFeatures while it's still running which means it didn't actually find the bestFeature to remove/add with best accuracy. Also added ina  copy() so that the temp array doesn't affect the original one